requirements.txt
django-registration
Django==1.10.3
Pillow==3.4.2
